requirements.txt
django==4.2
django-filter==25.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
django-modeltranslation==0.19.14
drf-spectacular==0.28.0
